unused extract_access_token
In the get_token method of the oauth2 gem, you can specify the access token class as the third argument. However, this is now deprecated, and if specified, you must implement the call method. To switch to the conventional access token, it is necessary to specify it in options[:access_token_class].

- https://gitlab.com/oauth-xx/oauth2/-/blob/v2.0.9/lib/oauth2/client.rb?ref_type=tags#L157